% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{DtBind}
\alias{DtBind}
\title{Bind Data}
\usage{
DtBind(MT)
}
\arguments{
\item{MT}{MT object}
}
\value{
Updated MT object with outcome data bound
}
\description{
Bind reshaped single-cell data to outcome data
}
