% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{DtReshape}
\alias{DtReshape}
\title{Reshape Data}
\usage{
DtReshape(MT, num_quan = 100, q.cut = 0, heterog = TRUE, prevelen = TRUE)
}
\arguments{
\item{MT}{MT object}

\item{num_quan}{Number of quantiles (default: 100)}

\item{q.cut}{Quantile cutoff threshold (default: 0)}

\item{heterog}{Whether to calculate heterogeneity (default: TRUE)}

\item{prevelen}{Whether to calculate prevalence (default: TRUE)}
}
\value{
Updated MT object with reshaped data
}
\description{
Reshape cell-level data to sample level features
}
