% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{HCalC}
\alias{HCalC}
\title{Calculate Heterogeneity}
\usage{
HCalC(dt_vector)
}
\arguments{
\item{dt_vector}{Numeric vector of element values}
}
\value{
Heterogeneity as double
}
\description{
Computes heterogeneity using C++ implementation for efficiency
}
\examples{
testdata <- c(1.2, 3.4, 2.1, 5.6)
h_val <- HCalC(testdata)
}
