% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R
\name{LoadSCDT}
\alias{LoadSCDT}
\title{Load Single Cell Data}
\usage{
LoadSCDT(MT)
}
\arguments{
\item{MT}{MT}
}
\value{
Updated MT object with loaded cell data
}
\description{
Loading single-cell data to MT
}
