% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{PerctCal}
\alias{PerctCal}
\title{Calculate Prevalence Percentage}
\usage{
PerctCal(dt.vector)
}
\arguments{
\item{dt.vector}{Data vector}
}
\value{
Prevalence percentage
}
\description{
Calculate Prevalence Percentage
}
