% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qSCMA-package.R
\docType{package}
\name{qSCMA}
\alias{qSCMA-package}
\alias{qSCMA}
\title{qSCMA: Quantile Single Cell Metallomic Analysis}
\description{
A package for quantitative single cell meta-analysis including
heterogeneity calculation, feature transformation, and machine
learning-based feature identification.
}
\author{
\strong{Maintainer}: Guohuan Zhang \email{granhamzhang@foxmail.com}

}
