% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{QuanCal}
\alias{QuanCal}
\title{Quantile Calculation Function}
\usage{
QuanCal(filedir, num_quan, q.cut, heterog, prevelen)
}
\arguments{
\item{filedir}{File directory}

\item{num_quan}{Number of quantiles}

\item{q.cut}{Quantile cutoff}

\item{heterog}{Whether to calculate heterogeneity}

\item{prevelen}{Whether to calculate prevalence}
}
\value{
Reshaped data with quantile features
}
\description{
Quantile Calculation Function
}
